Create Collect_Images.py
This code allows you to create your own dataset to train the model. When running the code, position your hand in front of the camera so that it is visible and press "q" to start recording. Move your hand around keeping the same gesture in order to create some variety on the dataset. 

Each frame will be saved as a separate image, the default setting is for 3 different hand gestures with 100 images each. you can change those parameters to your liking in the code.