Create Create_Dataset.py
This code should be executed after colleting the images for your dataset using the "Collect_Images.py" code. Here the images will be iterated and mediapipe will only collect the relevant information about the hands' landmarks. The landmarks represent the joints in your hand, and they are saved only as a coordinate on the screen.

The data of every landmark is saved into an array, and this is the dataset that we will be using to train the model (the position of the landmarks).

If the code is executed succesfully, there should appear a file in your directory called "data.pickle"